debug

+++ Datasets/Suites/Classify.py
@@ -74,7 +74,7 @@ class ClassifyEnv:
         if self.last:
             self.time_step = self.time_step._replace(step_type=StepType.LAST)
         else:
-            reward = self.time_step.label == np.expand_dims(np.argmax(action, -1), 1)
+            reward = (self.time_step.label == np.expand_dims(np.argmax(action, -1), 1)).astype('float32')
             self.time_step = self.time_step._replace(step_type=StepType.MID, reward=reward,
                                                      action=action)
         self.last = not self.last
